try fix CI
